direct_commands: split monolith into per-command-group package (#494)
The single 2510-line direct_commands.py is now a package directory
of 11 focused modules:

  __init__.py        public API + re-exports for backward compat
  _helpers.py        env / compose / k8s / SSH / registry plumbing
  info.py            list, version, status (~400 lines)
  lifecycle.py       start, stop, restart, scale (~245 lines)
  config.py          config get
  maintenance.py     maintenance script/extension/update
  host.py            host list/add/remove
  gitops.py          gitops status/diff (~307 lines)
  extension_skin.py  extension/skin list
  backup.py          backup list
  doctor.py          doctor (~205 lines)
  argocd.py          argocd password/apps/ui

Sub-modules call shared helpers through `_helpers.foo()` rather than
local-imported names, so test patches at
`direct_commands._helpers.<helper>` propagate to every caller.

`__init__.py` re-exports every previously-module-level symbol the
test suite reaches for so existing `direct_commands.X` access keeps
working. Side-effect imports of each handler module ensure the
@register decorators populate DIRECT_COMMANDS at package import.

Test patches were updated mechanically: occurrences of
`monkeypatch.setattr(direct_commands, "_helper", …)` became
`monkeypatch.setattr(direct_commands._helpers, "_helper", …)`,
plus three patches retargeted to the per-command sub-module that
actually owns the helper (_argocd_admin_password → argocd,
_read_wiki_ids → maintenance). One small move: _k8s_namespace
moved from lifecycle to _helpers because _exec_in_container needs
it.

No behavior change. 709 unit tests pass (211 in
test_direct_commands.py).

Closes #494.
